requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.18
python-dotenv>=1.0.0
azure-identity>=1.15.0
azure-search-documents>=11.4.0
tiktoken>=0.5.1
requests